+++ llvm/include/llvm/Support/ELFAttrParserExtended.h
@@ -39,8 +39,8 @@ public:
   ~ELFExtendedAttrParser() override { static_cast<void>(!Cursor.takeError()); }
   Error parse(ArrayRef<uint8_t> Section, llvm::endianness Endian) override;
 
-  std::optional<unsigned> getAttributeValue(unsigned Tag) const override;
-  std::optional<unsigned> getAttributeValue(StringRef BuildAttrSubsectionName,
+  std::optional<uint64_t> getAttributeValue(unsigned Tag) const override;
+  std::optional<uint64_t> getAttributeValue(StringRef BuildAttrSubsectionName,
                                             unsigned Tag) const override;
   std::optional<StringRef> getAttributeString(unsigned Tag) const override;
   std::optional<StringRef> getAttributeString(StringRef BuildAttrSubsectionName,
